# src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description: 智慧停车应用
  version: 1.0.0
  title: 智慧停车应用
basePath: "/"
schemes:
- http

##认证
#securityDefinitions:
#  api_key:
#    type: "apiKey"
#    name: "token"
#    in: "header"

#定义接口数据
paths:
  /merchant/car/paylist:
    get:
      tags:
      - 停车统计
      summary: 停车统计列表
      produces:
      - application/json
      parameters:
      - name: page
        description: 页码
        in: query
        type: integer
        required: true
      - name: limit
        description: 每页条数
        in: query
        type: integer
        required: true
      - name: sidx
        description: 排序字段
        in: query
        type: string
      - name: plateNo
        description: 车牌号模糊查询
        in: query
        type: string
      - name: payTimeStart
        description: 支付起始日期模糊查询
        in: query
        type: string
        format: date-time
      - name: payTimeEnd
        description: 支付结束日期模糊查询
        in: query
        type: string
        format: date-time
      responses:
        '200':
          description: 返回车牌缴费列表
          schema:
            type: array
            items:
              $ref: '#/definitions/PayOrderInfo'
  /merchant/car/payinfo/{id}:
    get:
      tags:
      - 停车统计
      summary: 停车缴费详情
      produces:
      - application/json
      parameters:
      - name: id
        description: 缴费明细id
        in: path
        type: String
        required: true
      responses:
        '200':
          description: 返回停车缴费详情
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              data:
                description: 缴费详情
                $ref: '#/definitions/PayOrderInfo'
  /merchant/car/list:
    get:
      tags:
      - 停车统计
      summary: 车牌统计列表查询
      produces:
      - application/json
      parameters:
      - name: page
        description: 页码
        in: query
        type: integer
        required: true
      - name: limit
        description: 每页条数
        in: query
        type: integer
        required: true
      - name: sidx
        description: 排序字段
        in: query
        type: string
      - name: plateNo
        description: 车牌号模糊查询
        in: query
        type: string
      responses:
          '200':
            description: 返回车牌统计列表
            schema:
              type: array
              items:
                $ref: '#/definitions/PlateNoInfo'
  /merchant/car/infolist:
    get:
      tags:
      - 停车统计
      summary: 车牌所有缴费列表查询
      produces:
      - application/json
      parameters:
      - name: page
        description: 页码
        in: query
        type: integer
        required: true
      - name: limit
        description: 每页条数
        in: query
        type: integer
        required: true
      - name: sidx
        description: 排序字段
        in: query
        type: string
      - name: plateNo
        description: 车牌号查询
        in: query
        type: string
      responses:
        '200':
          description: 返回停车统计缴费详情
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              data:
                description: 菜单信息
                $ref: '#/definitions/PayOrderInfo'
...
  /merchant/Advertisement/info/{id}:
    get:
      tags:
      - 营销管理
      summary: 获取广告模板详情
      produces:
      - application/json
      parameters:
      - name: id
        description: 广告模板id
        in: path
        type: String
        required: true
      responses:
        '200':
          description: 返回广告模板信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              data:
                description: 菜单信息
                $ref: '#/definitions/Advertisement'
  /merchant/Advertisement/add:
    post:
      tags:
      - 营销管理
      summary: 添加广告模板
      produces:
      - application/json
      parameters:
      - name: body
        description: 广告模板对象
        in: body
        type: string
        schema:
          $ref: '#/definitions/AdvertisementSave'
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /merchant/Advertisement/update:
    patch:
      tags:
      - 营销管理
      summary: 修改广告模板
      produces:
      - application/json
      parameters:
      - name: body
        description: 广告模板对象
        in: body
        type: string
        schema:
          $ref: '#/definitions/AdvertisementEdit'
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /merchant/Advertisement/delete/{id}:
    delete:
      tags:
      - 营销管理
      summary: 广告模板删除
      produces:
      - application/json
      parameters:
      - name: id
        description: 广告模板id
        in: path
        type: String
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
  /user/order/payinfo:
    get:
      tags:
      - 用户接口
      summary: 停车缴费费用查询
      produces:
      - application/json
      parameters:
      - name: body
        description: 车牌号和车场信息
        in: body
        type: string
        schema:
          $ref: '#/definitions/CarInfo'
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /user/order/pay:
    post:
      tags:
      - 用户接口
      summary: 停车费用支付
      produces:
      - application/json
      parameters:
      - name: body
        description: 停车费用参数
        in: body
        type: string
        schema:
          $ref: '#/definitions/PayOrderInfo'
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
definitions:
  R:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      msg:
        description: 失败原因
        type: string
...
  PlateNoInfo:
    type: object
    properties:
      id:
        description: 主键id
        type: string
      merId:
        description: 商户id
        type: string
      openId:
        description: openId
        type: string
      nickName:
        description: 昵称
        type: string
      plateNo:
        description: 车牌号
        type: string
...
  PayOrderInfo:
    type: object
    properties:
      id:
        description: 主键id
        type: string
      merId:
        description: 商户id
        type: string
      openId:
        description: openId
        type: string
      nickName:
        description: 昵称
        type: string
      plateNo:
        description: 车牌号
        type: string
      parkId:
        description: 停车场id
        type: integer
        format: int64
      orderNo:
        description: 订单号
        type: string
      parkName:
        description: 停车场名称
        type: string
      entryTime:
        description: 入场时间
        type: string
        format: date-time
      payTime:
        description: 支付时间
        type: string
        format: date-time
      elapsedTime:
        description: 停留时间
        type: integer
        format: int64
      img:
        description: 停车图片
        type: string
      payable:
        description: 支付金额
        type: integer
      delayTime:
        description: 延迟出场时间
        type: integer
...
  Advertisement:
    type: object
    properties:
      id:
        description: 广告模板id
        type: string
      templateId:
        description: 广告模板类型id
        type: string
      tempCategoryCode:
        description: 广告模板类型代码
        type: string
      templateNo:
        description: 广告模板代码
        type: string
      templateName:
        description: 广告模板名称
        type: string
      templateUrl:
        description: 广告模板url
        type: string
      templateImg:
        description: 广告模板图片
        type: string
      templateStatus:
        description: 广告模板状态
        type: integer
        format: int64
      tempCategoryDesc:
        description: 广告模板类型名称
        type: string
      note:
        description: 广告模板备注
        type: string
      createBy:
        description: 广告创建人
        type: integer
        format: int64
      createTime:
        description: 广告创建时间
        type: string
        format: date-time
      lastUpdatedBy:
        description: 广告最后更新人
        type: integer
        format: int64
      lastUpdatedTime:
        description: 广告最后更新时间
        type: string
        format: date-time
  AdvertisementSave:
    type: object
    properties:
      templateId:
          description: 广告模板类型id
          type: string
      templateName:
          description: 广告模板名称
          type: string
      templateUrl:
          description: 广告模板url
          type: string
      templateImg:
          description: 广告模板图片
          type: string
          format: string
      tempCategoryCode:
          description: 广告类型代码
          type: string
      note:
          description: 广告模板备注
          type: string
  AdvertisementEdit:
    type: object
    properties:
      id:
        description: 广告模板id
        type: string
      templateId:
        description: 广告模板类型id
        type: string
      templateName:
        description: 广告模板名称
        type: string
      templateUrl:
        description: 广告模板url
        type: string
      templateImg:
        description: 广告模板图片
        type: string
      templateStatus:
        description: 广告模板状态
        type: integer
        format: int64
      note:
        description: 广告模板备注
        type: string
...
  CarInfo:
    type: object
    properties:
      plateNo:
        description: 广告类型名称
        type: string
      parkId:
        description: 广告类型状态
        type: integer
        format: int64
...